Fix region config defaults + use region id if description is missing.
In case the region description was missing, EU868 was set as default. As
well the default configuration (in case no regions were set) would
always fallback to EU868 which was used for testing. This removes the
default EU868 configuration and moves this configuration to tests.

This updates the region API methods to use the region id as description,
in case the description is not configured.

Closes #120.