Make it possible to configure failover mqtt servers
This adds a "servers" setting, removing the "server" one, making it
possible to configure fail-over MQTT servers for use with HA MQTT
configurations.

To avoid confusion around what happens when "server" is set and
"servers" is not, vs. what happens when server is not part of servers,
this removes the old "server" setting and forces it to be a single array
of servers, similar to how Redis configuration is done.

This means that upgrading may require a configuration change.

Signed-off-by: D.S. Ljungmark <spider@skuggor.se>