ui: Make it more clear if fuota job is running or scheduled.
Before, the spinner icon would show for both running jobs + jobs
that were scheduled in the future. With this change, a clock icon
is showed when a job is scheduled in the future which which gets changed
into the spinner icon once it will get executed. This also add a column
displaying the run at timestamp.