Move f_cnt_up out of device-session into device record.
By moving out the f_cnt_up and moving it into the device record, we
can remove the transaction and record locking (using for update) from
the deduplication logic. Instead of locking it, we can increment the
f_cnt_up like:

  update device
  set f_cnt_up = NEW_FCNT
  where
    dev_eui = DEV_EUI and
    f_cnt_up = OLD_FCNT

If that query results in 0 records being updated, we know there is a
race-condition, meaning that there is already an other process handling
the uplink, and we should abort.

Removing the transaction / locking records also solves a rare deadlock
as the deduplication query filtering on dev_addr might lock more records
than needed.