lrwn: Add get_new_channel_req_dr_range fn.
This function given the data-rates supported by the device, returns the
min_dr / max_dr values to be used with the NewChannelReq mac-command.
Note that in some cases (e.g. in EU868) the min_dr / max_dr have a
special meaning (in which case the max_dr is 0, and the min_dr > 0).