Refactor device-profile templates into device profiles.
This refactors the device-profile templates, which could be copied by a
tenant into a device profile, into just device profiles. By making the
foreign-key to tenant_id nullable, there can be global device profiles
and profiles owned by a tenant.

This also adds device_profile_device and device_profile_vendor tables,
such that we can store data related to a vendor and devices in an
organized way.

Existing device-profile templates can be migrated into device profiles
using the `migrate-device-profile-templates` CLI command.

While this commit removes the device-profile templates from the code,
API and UI, it does not (yet) remove the templates from the database.
This table will be removed in the future.