Refactor datarate ranges / support RP002-1.0.5 in NewChannelReq.
This refactors the min_dr / max_dr ranges into datarates, as with
RP002-1.0.5, there are ranges where min_dr > max_dr (e.g. to encode
DR0-DR5 + DR12 + DR13).

This also adds a new field supported_uplink_data_rates to the
device-profile, such that it is possible to configure which datarates
are supported by the end-device. In the NewChannelReq, ChirpStack will
only include the datarates that are supported by both the network and
the end-device.